+++ cogs/starboard.py
@@ -1,27 +1,25 @@
 import logging
-import sys
 from datetime import *
 
 import discord
-from discord import RawReactionActionEvent, errors
+from discord import RawReactionActionEvent
 from discord.ext import commands, tasks
+from discord.ext.commands import PartialEmojiConversionFailure
+from discord_slash import cog_ext, SlashContext
 
+from converters.converters import PartialEmojiWithUnicodeConverter
 from model.embeds import StarboardEmbed
 from model.model import *
-from .utilities import credential_checks
 
 logger = logging.getLogger('discord')
 
 
-def _get_emoji_for_star(stars):
-    if 5 >= stars >= 0:
-        return '\N{WHITE MEDIUM STAR}'
-    elif 10 >= stars >= 6:
-        return '\N{GLOWING STAR}'
-    elif 25 >= stars >= 11:
-        return '\N{DIZZY SYMBOL}'
-    else:
-        return '\N{SPARKLES}'
+async def _delete_starred_message(message_to_delete: StarredMessageModel, starboard_message: discord.Message):
+    message_to_delete.embed_message_id = None
+    message_to_delete.save()
+
+    if starboard_message is not None:
+        await starboard_message.delete()
 
 
 class Starboard(commands.Cog):
@@ -39,77 +37,155 @@ class Starboard(commands.Cog):
 
     @tasks.loop(minutes=5)
     async def cleaner(self):
-        # The cleaner checks all starboard messages in the past week.
-        threshold = datetime.today() - timedelta(days=1)
-        messages_to_check = StarredMessageModel.select().where(StarredMessageModel.datetime_added > threshold)
+        """ A process which runs every 5 minutes to ensure all messages are removed if necessary. """
+        cleaning_threshold = datetime.today() - timedelta(days=1)
+        messages_to_check = StarredMessageModel.select().where(StarredMessageModel.datetime_added > cleaning_threshold)
+
+        logger.info("Running the cleaner for {} starboard message(s).".format(len(messages_to_check)))
 
         # For each message, it updates the embed.
         for message_to_check in messages_to_check:
-            channel = self.client.get_channel(message_to_check.message_channel_id)
+            original_message_channel = self.client.get_channel(message_to_check.message_channel_id)
+            starboard_channel = self.client.get_channel(message_to_check.starboard.channel_id)
+            starboard_message = None
 
-            # The bot may have been removed from the server.
-            if channel is None:
+            # The bot may have been removed from the server, in which case skip it for now.
+            if original_message_channel is None:
                 continue
 
             try:
-                discord_message = await channel.fetch_message(message_to_check.message_id)
-                embed = await self._get_starred_embed(message_to_check, discord_message, True)
-            except:
-                # Temporary error handling until discord.py releases 1.4.
-                e = sys.exc_info()[0]
-                logger.error("Error processing starboard messages: {}".format(e))
-                continue
+                original_message = await original_message_channel.fetch_message(message_to_check.message_id)
+                new_embed = await self._get_starred_embed(message_to_check, original_message)
 
-            await self._update_starred_message(message_to_check, embed)
-            logger.info("Checked {.message_id}".format(message_to_check))
+                starboard_message = await starboard_channel.fetch_message(message_to_check.embed_message_id)
+            except (discord.NotFound, discord.Forbidden, discord.HTTPException):
+                new_embed = None
 
-            # Only check if the message meets the threshold when cleaning up.
-            embed = await self._get_starred_embed(message_to_check, discord_message, True)
-            await self._update_starred_message(message_to_check, embed)
+            star_threshold = message_to_check.starboard.star_threshold
+            number_of_stars = len(message_to_check.starrers)
+
+            if message_to_check.starboard.star_threshold != 1 and number_of_stars < star_threshold:
+                await _delete_starred_message(message_to_check, starboard_message)
+            else:
+                await self._update_starred_message(message_to_check, new_embed)
 
     @cleaner.before_loop
     async def before_cleaner(self):
         await self.client.wait_until_ready()
 
-    @commands.command(no_pm=True)
-    @credential_checks.has_permissions(administrator=True)
-    async def starboard(self, ctx, *, channel: discord.TextChannel = None):
-        """Sets up the starboard for this server.
-        This creates a starboard in the specified channel
-        this makes it into the server's "starboard".
+    async def _get_starred_embed(self, starred_message: StarredMessageModel, discord_message: discord.Message):
+        """ Gets an embed when provided with a starred message and the message that should be embedded. """
+        # Check the number of stars, if the message has zero stars then we no longer render it.
+        number_of_stars = len(starred_message.starrers)
+        star_emoji = starred_message.starboard.emoji
 
-        If the channel is deleted then the starboard is
-        deleted as well.
-
-        You must have Administrator permissions to use this
-        command.
-        """
-
-        if channel is None:
-            ctx.send("You need to provide an existing channel to turn it into a Starboard!")
+        if number_of_stars == 0:
             return
 
-        # Check the database for a starboard.
-        previous_starboard = StarboardModel.get_for_guild(channel.guild.id)
+        # Otherwise return a stylised embed for the message.
+        e = StarboardEmbed(message=discord_message,
+                           starred_message=starred_message,
+                           cleaner_next_iteration=self.cleaner.next_iteration,
+                           star_emoji=star_emoji)
 
-        if previous_starboard is not None:
-            # Then check if the channel still exists.
-            previous_channel = self.client.get_channel(previous_starboard.channel_id)
-            if previous_channel is not None:
-                fail_message = 'This server already has a starboard ({.mention})'
-                await ctx.send(fail_message.format(previous_channel))
-                return
-            else:
-                clear_message = "This server already _had_ a starboard. I notice it no longer exists or I can't get " \
-                                "it. I will delete the old record."
+        return e.populate()
 
-                await ctx.send(clear_message)
-                previous_starboard.delete_instance()
+    async def _on_reaction_removed(self, reaction: RawReactionActionEvent):
+        starboard = StarboardModel.get_for_guild(reaction.guild_id, reaction.emoji)
+        starred_message = await StarredMessageModel.get_in_starboard(reaction.message_id, starboard.id)
 
-        bot_permissions = ctx.message.channel.permissions_for(ctx.guild.me)
-        await ctx.send("Updating permissions for {.mention}".format(channel))
+        if starred_message is None:
+            return
+
+        await MessageStarrerModel.delete_starrer_message(message_id=starred_message.id, starrer_id=reaction.user_id)
+
+        message_channel = self.client.get_channel(reaction.channel_id)
+        unstarred_message = await message_channel.fetch_message(reaction.message_id)
+
+        new_embed = await self._get_starred_embed(starred_message, unstarred_message)
+
+        await self._update_starred_message(starred_message, new_embed)
+
+    async def _on_reaction(self, reaction: RawReactionActionEvent):
+        starboard = StarboardModel.get_for_guild(reaction.guild_id, reaction.emoji)
+
+        if starboard is None:
+            return
+
+        channel = self.client.get_channel(reaction.channel_id)
+        starred_message = await StarredMessageModel.get_in_starboard(reaction.message_id, starboard.id)
+        discord_message = await channel.fetch_message(reaction.message_id)
+
+        # Users cannot star their own message.
+        if discord_message.author.id == reaction.user_id:
+            return
+
+        if starred_message is None:
+            starred_message = await StarredMessageModel.create_from_message(message=discord_message, starboard=starboard)
+
+        if starred_message.is_muted:
+            return
+
+        await MessageStarrerModel.add_starred_message(message_id=starred_message.id, starrer_id=reaction.user_id)
+
+        # Show it in the starboard.
+        embed = await self._get_starred_embed(starred_message, discord_message)
+        await self._update_starred_message(starred_message, embed)
+
+    async def _update_starred_message(self, starred_message: StarredMessageModel, new_embed: discord.Embed):
+        """ Updates a starred message in Discord.
+        If the embed provided is None then it attempts to delete the message. """
+
+        starboard_channel = self.client.get_channel(starred_message.starboard.channel_id)
+
+        # If this is the first time submitting this embed, post a new message for it.
+        if starred_message.embed_message_id is None and new_embed is not None:
+            embed_message = await starboard_channel.send(embed=new_embed)
+
+            starred_message.embed_message_id = embed_message.id
+            starred_message.save()
+
+            return
+
+        # Otherwise, we look for the message we posted previously and attempt to update it.
+        try:
+            original_message = await starboard_channel.fetch_message(starred_message.embed_message_id)
+        except (discord.NotFound, discord.HTTPException):
+            return
+
+        # If we have been provided an empty embed then it needs deleting.
+        if new_embed is None:
+            return await _delete_starred_message(starred_message, original_message)
+
+        # Finally, render the new embed.
+        await original_message.edit(embed=new_embed)
+
+    @cog_ext.cog_subcommand(base="starboard", name="setup",
+                            description="Sets up a starboard for a given channel - updates it if there already is one in the channel.",
+                            options=[{"name": "channel", "description": "The channel to set the class up in.", "type": 7, "required": True},
+                                     {"name": "emoji", "description": "The emoji for the starboard to use.", "type": 3, "required": True},
+                                     {"name": "threshold", "description": "If a message has under these many stars, it will be automatically deleted.", "type": 4}],
+                            guild_ids=[197972184466063381])
+    @commands.has_permissions(manage_guild=True)
+    async def _setup_starboard(self, ctx: SlashContext, channel: discord.TextChannel, emoji: str, threshold: int = 1):
+        if not isinstance(channel, discord.TextChannel):
+            return await ctx.send("The channel you have provided is not supported.")
+
+        emoji_converter = PartialEmojiWithUnicodeConverter()
+
+        try:
+            emoji = await emoji_converter.convert(ctx, emoji)
+        except PartialEmojiConversionFailure:
+            return await ctx.send("Unknown emoji: {}".format(emoji))
+
+        if emoji not in ctx.guild.emojis and not isinstance(emoji, str):
+            await ctx.send("Warning, the emoji you have picked doesn't exist in this server. It could be deleted "
+                           "without your knowledge!")
+
+        StarboardModel.add_or_update(ctx.guild.id, channel.id, str(emoji), threshold)
 
         args = []
+        bot_permissions = ctx.channel.permissions_for(ctx.guild.me)
 
         # Make sure that people cannot send messages in the starboard.
         if bot_permissions.manage_roles:
@@ -121,160 +197,7 @@ class Starboard(commands.Cog):
         else:
             await ctx.send("I can't update the channel permissions to stop people talking. Maybe check that?")
 
-        StarboardModel.create(guild_id=ctx.guild.id, channel_id=channel.id, is_locked=False)
-        await ctx.send('\N{GLOWING STAR} Starboard set up at {.mention}'.format(channel))
-
-    @commands.command(no_pm=True, aliases=['threshold'])
-    @credential_checks.has_permissions(administrator=True)
-    async def starboardthreshold(self, ctx, threshold: int):
-        """
-        A command which allows you to set the amount of stars a single message can receive before it makes it into the starboard.
-
-        You must have administrator permissions to perform this.
-        """
-        starboard = StarboardModel.get_for_guild(ctx.guild.id)
-
-        if starboard is None:
-            return await ctx.send(
-                "Couldn't find a starboard for this server. Use the `!starboard` command to set one up.")
-
-        if threshold < 0:
-            return await ctx.send("You can't have negative stars!")
-
-        starboard.star_threshold = threshold
-        starboard.save()
-
-        await ctx.send(
-            "Ok! Your starboard's threshold has been set to {}. Messages within the past week will be updated soon.".format(
-                threshold))
-
-    async def _get_emoji(self, guild_id):
-        guild_config = await GuildConfig.get_for_guild(guild_id)
-
-        if guild_config.starboard_emoji_id is None:
-            return "⭐"
-
-        return self.client.get_emoji(guild_config.starboard_emoji_id)
-
-    async def _get_starred_embed(self, starred_message: StarredMessageModel, discord_message: discord.Message,
-                                 remove_after_threshold: bool = False):
-
-        number_of_stars = len(starred_message.starrers)
-        star_emoji = await self._get_emoji(discord_message.guild.id)
-
-        if number_of_stars == 0:
-            return
-
-        e = StarboardEmbed(message=discord_message,
-                           starred_message=starred_message,
-                           remove_after_threshold=remove_after_threshold,
-                           cleaner_next_iteration=self.cleaner.next_iteration,
-                           star_emoji=star_emoji)
-
-        return e.populate()
-
-    async def _on_reaction_removed(self, reaction: RawReactionActionEvent):
-        guild_emoji = await self._get_emoji(reaction.guild_id)
-
-        if isinstance(guild_emoji, str):
-            if reaction.emoji.name != guild_emoji:
-                return
-
-        else:
-            if reaction.emoji.id != guild_emoji.id:
-                return
-
-        starred_message = StarredMessageModel.get_or_none(reaction.message_id)
-
-        if starred_message is None:
-            return
-
-        MessageStarrerModel.delete().where((MessageStarrerModel.message_id == reaction.message_id) &
-                                           (MessageStarrerModel.user_id == reaction.user_id)).execute()
-
-        channel = self.client.get_channel(starred_message.starboard.channel_id)
-        discord_message = await channel.fetch_message(reaction.message_id)
-
-        embed = await self._get_starred_embed(starred_message, discord_message)
-
-        await self._update_starred_message(starred_message, embed)
-
-    async def _on_reaction(self, reaction: RawReactionActionEvent):
-        guild_emoji = await self._get_emoji(reaction.guild_id)
-
-        if isinstance(guild_emoji, str):
-            if reaction.emoji.name != guild_emoji:
-                return
-
-        else:
-            if reaction.emoji.id != guild_emoji.id:
-                return
-
-        starboard = StarboardModel.get_for_guild(reaction.guild_id)
-
-        if starboard is None:
-            return
-
-        channel = self.client.get_channel(reaction.channel_id)
-        starred_message = StarredMessageModel.get_or_none(reaction.message_id)
-        discord_message = await channel.fetch_message(reaction.message_id)
-
-        if starred_message is None:
-            starred_message = StarredMessageModel.create(message_id=discord_message.id,
-                                                         message_channel_id=discord_message.channel.id,
-                                                         starboard_id=starboard.id,
-                                                         is_muted=False, datetime_added=datetime.utcnow(),
-                                                         user_id=discord_message.author.id)
-
-        # Don't interact with muted messages or let people star themselves.
-        if starred_message.is_muted or discord_message.author.id == reaction.user_id:
-            return
-
-        # Add the reactor.
-        insert = MessageStarrerModel.insert(message_id=starred_message.message_id, user_id=reaction.user_id,
-                                            datetime_starred=datetime.utcnow()) \
-            .on_conflict(
-            update={MessageStarrerModel.datetime_starred: datetime.utcnow()}) \
-            .execute()
-
-        # Show it in the starboard.
-        embed = await self._get_starred_embed(starred_message, discord_message)
-        await self._update_starred_message(starred_message, embed)
-
-    async def _get_guild_star(self, guild_id: int):
-        pass
-
-    async def _update_starred_message(self, starred_message: StarredMessageModel,
-                                      embed: discord.Embed):
-
-        starboard_channel = self.client.get_channel(starred_message.starboard.channel_id)
-
-        # Either send it (if it hasn't already been submitted to the starboard) or update it.
-        if starred_message.embed_message_id is None:
-            # If we don't have a message or an embed, continue the logic. We can sort it later.
-            if embed is None:
-                return
-
-            embed_message = await starboard_channel.send(embed=embed)
-
-            starred_message.embed_message_id = embed_message.id
-            starred_message.save()
-        else:
-            try:
-                embed_message = await starboard_channel.fetch_message(starred_message.embed_message_id)
-            except errors.NotFound:
-                return
-
-            if embed is not None:
-                await embed_message.edit(embed=embed)
-                return
-
-            # If the embed can no longer be generated but a message has been posted for it:
-            # Remove the embed ID and delete the message.
-            starred_message.embed_message_id = None
-            starred_message.save()
-
-            return await embed_message.delete()
+        await ctx.send('{} Starboard set up at {.mention}. Permissions have been updated.'.format(emoji, channel))
 
 
 def setup(client):