+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Slash Commands
+- Starboards can now be configured using slash commands. The original way of adding starboards has been removed.
+- Server configuration can be modified using slash commands.
+
+### Features
+- Servers can now have more than one starboard. 
+- Custom starboard emoji can now also include the default Discord unicode emoji.
+- Greatly streamlined the starboard process.
+
+### Maintenance
+- Version number is now automatically picked up in the CI process.
 
 ## [4.7.4] - 2021-03-31
 ### Tweaks